Move multilabel end-to-end Datlab tests to test_multilabel.py module within datalab test suite
